another attempt making dosobjs 32bit

+++ include/dj64/thunks_p.h
@@ -25,9 +25,9 @@ uint64_t dj64_asm_call(int num, uint8_t *sp, uint8_t len,
 uint64_t dj64_asm_call_u(int handle, int libid, int num, uint8_t *sp,
     uint8_t len, int flags);
 uint8_t *dj64_clean_stk(size_t len);
-uint32_t dj64_obj_init(const void *data, uint16_t len);
-uint32_t dj64_obj_oinit(const void *data, uint16_t len);
-void dj64_obj_done(void *data, uint32_t fa, uint16_t len);
+uint32_t dj64_obj_init(const void *data, uint32_t len);
+uint32_t dj64_obj_oinit(const void *data, uint32_t len);
+void dj64_obj_done(void *data, uint32_t fa, uint32_t len);
 void dj64_rm_dosobj(const void *data, uint32_t fa);
 
 #define _TFLG_NONE 0
